fix async issue in text display
Gui.cs is apparently doing it's own async message pumping, and the messages only get pumped when there's some external event (e.g. mouse moving over the console window). To fix this, we use Task.Run to move the task to the thread pool, and the `Application.MainLoop.Invoke` API to move the screen updates back to the UI thread.